refactor(Sheet): separated sheet hash logic
